add tests for includes
